Rename RoPE.py to embedding.py. We are not going to use RoPE in this project.
